Feat: refact dashboard data structure
